requirements.txt
pyTelegramBotAPI==4.8.0
flake8==5.0.4
python-dotenv==0.21.0
requests==2.28.1
lxml==4.9.1
pandas==1.3.5
openpyxl==3.0.10
xlrd==2.0.1
selenium==4.7.2
